# openapi.yaml
openapi: '3.0.3'
info:
  title: API para eventos
  version: '0.1.0'
  description: |
    # Título
    API para eventos que permite operaciones CRUD con eventos, lugares y organizadores.
    
    ## Subtítulo
    Ejemplo
...
paths:
...
  /organizadores:
    get:
      summary: Obtener todos los organizadores
      description:
        Obtener un listado de todos los organizadores existentes
      operationId: getOrganizadores
      tags:
        - organizador
      responses:
        '200':
          description: OK
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/organizadores'
    post:
      summary: Crear un nuevo organizador
      description: TODO
      operationId: createOrganizador
      tags:
        - organizador
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/organizador'
      responses:
        '201':
          description: Organizador creado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/info_nuevo_elemento'
        default:
          description: Error al crear el organizador
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'  
  /organizadores/{id}:
    get:
      summary: Obtener información de un organizador concreto
      description: TODO
      operationId: getOrganizador
      tags:
        - organizador
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/organizador'
            application/xml:
              schema:
                $ref: '#/components/schemas/organizador'
        default:
          description: Error al acceder al organizador
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
      parameters:
        - $ref: '#/components/parameters/id'
    put:
      summary: Modificar información de un organizador concreto
      description: TODO
      operationId: updateOrganizador
      tags:
        - organizador
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '201':
          description: Ok
        default:
          description: Error al acceder al organizador
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/organizador_editado'
    delete:
      summary: Borrar un organizador especifico
      description: TODO
      operationId: deleteOrganizador
      tags:
        - organizador
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Ok
        default:
          description: Error al acceder al organizador
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
components:
  parameters:
    id:
      name: id
      in: path
      description: ID del elemento
      required: true
      schema:
        $ref: '#/components/schemas/id'
  schemas:
    id:
      type: integer
      minimum: 1
    error:
      type: object
      properties:
        descripcion:
          type: string
        codigo:
          type: integer
      example: {
        "descripcion": "Datos introducidos erroneos",
        "codigo": 400}
    info_nuevo_elemento:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/id'
      example: {"id": 123}
...
    organizadores:
      type: array
      items:
        $ref: '#/components/schemas/organizador'
...
    organizador:
      allOf:
        - $ref: '#/components/schemas/organizador_editado'
        - required:
          - nombre_empresa
          - pais_origen
        - oneOf:
          - required:
            - owner
          - required:
            - owners
      example:
        {"nombre_empresa": "De la Riva",
        "pais_origen": "España",
        "owner": "José Morán Arteaga"
        }
...
    organizador_editado:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        nombre_empresa:
          type: string
        pais_origen:
          type: string
        owner:
          $ref: '#/components/schemas/persona'
        owners:
          $ref: '#/components/schemas/array_personas'
      example:
        {"nombre": "Santiago Bernabeu"}
...
    array_personas:
      type: array
      items:
        $ref: '#/components/schemas/persona'
    persona:
      type: object
      required:
        - nombre
      properties:
        nombre:
          type: string
        apellidos:
          type: string
        edad:
          type: integer
      example: {"nombre": "David", "apellidos": "Ashby"}